refactor(domain): extract scoring localization helpers
